refactor: Refactor controller directory
Files in controller directory are seperated to remove complexity
and related routers are updated.
