+++ server/controller/functions/encryption.js
