+++ server/controller/functions/token.js
