+++ server/controller/index.js
@@ -1,0 +1,19 @@
+module.exports = {
+  signup: require("./auth/signup"),
+  signin: require("./auth/signin"),
+  checkAvailability: require("./auth/checkAvailability"),
+  signout: require("./auth/signout"),
+  withdraw: require("./auth/withdraw"),
+  refreshToken: require("./auth/refresh"),
+  googleSignin: require("./oauth/googleSignin"),
+  kakaoSignin: require("./oauth/kakaoSignin"),
+  updateStudylog: require("./studylog/updateStudyLog"),
+  getStudylog: require("./studylog/getStudyLog"),
+  comment: require("./studylog/comment"),
+  studyToggle: require("./studylog/studyToggle"),
+  userInfo: require("./user/userInfo"),
+  profile: require("./user/profile"),
+  follows: require("./user/follows"),
+  achievement: require("./user/achievement"),
+  report: require("./user/report"),
+};
