+++ server/controller/studylog/updateStudyLog.js
@@ -1,12 +1,8 @@
-const { Studylog, Currentlog } = require("../models");
-const { isAccessAuthorized } = require("./functions/tokenFunc");
-const {
-  getStudyTime,
-  getStudyLogs,
-  toggleOff,
-} = require("./functions/modelFunc");
-const checkAchievement = require("./functions/checkAchievement");
 const { Op } = require("sequelize");
+const { Studylog, Currentlog } = require("../../models");
+const { isAccessAuthorized } = require("../functions/token");
+const { toggleOff } = require("../functions/model");
+const checkAchievement = require("../functions/achievementCheck");
 
 module.exports = {
   post: async (req, res) => {
@@ -86,45 +82,6 @@ module.exports = {
       } else {
         res.sendStatus(500);
       }
-    } catch (e) {
-      res.status(500).send(e);
-    }
-  },
-
-  get: async (req, res) => {
-    try {
-      const user = isAccessAuthorized(req);
-      const date =
-        req.query.date ||
-        new Date().getFullYear() +
-          new Date().getMonth() +
-          new Date().getDate() +
-          "";
-      const offset = req.query.offset || 0;
-
-      const utc =
-        Number(offset) +
-        Date.UTC(
-          date.slice(0, 4),
-          Number(date.slice(4, 6)) - 1,
-          Number(date.slice(6, 8))
-        ) /
-          (1000 * 60);
-      console.log("utc", utc);
-
-      const dateStart = Math.round(utc);
-      const dateEnd = dateStart + 24 * 60;
-
-      console.log(dateStart);
-      console.log(dateEnd);
-
-      const studyLogList = await getStudyLogs(user.id, dateStart, dateEnd);
-      const studyTime = await getStudyTime(studyLogList);
-
-      res.send({
-        studylogList: studyLogList,
-        studyTime: studyTime,
-      });
     } catch (e) {
       res.status(500).send(e);
     }
