+++ server/routes/verify.js
@@ -1,7 +1,7 @@
 const express = require("express");
 const router = express.Router();
-const verifyEmail = require("../controller/verifyEmail");
+const verifyEmail = require("../controller/auth/verifyEmail");
 
-router.get("/:emailVerifyCode", verifyEmail);
+router.post("/:emailVerifyCode", verifyEmail);
 
 module.exports = router;
